browser port lookup back to async
