+++ tiberius/src/client.rs
@@ -6,7 +6,6 @@ mod auth;
 pub use builder::*;
 pub use auth::*;
 pub(crate) use connection::*;
-pub use connection::find_tcp_port;
 
 use crate::{
     result::{ExecuteResult, QueryResult},
