+++ tiberius/src/lib.rs
@@ -18,7 +18,7 @@ mod result;
 mod row;
 mod tds;
 
-pub use client::{AuthMethod, Client, ClientBuilder, find_tcp_port};
+pub use client::{AuthMethod, Client, ClientBuilder};
 pub(crate) use error::Error;
 pub use from_sql::{FromSql, FromSqlOwned};
 pub use result::*;
